+++ src/SingleParameterFactory.php
@@ -17,15 +17,21 @@ interface SingleParameterFactory
     /**
      * @param TInput|static $value
      */
-    public static function instance(mixed $value): static;
+    public static function instance(
+        mixed $value
+    ): static;
 
     /**
      * @param TInput|static|null $value
      */
-    public static function orNull(mixed $value): ?static;
+    public static function orNull(
+        mixed $value
+    ): ?static;
 
     /**
      * @param TInput $value
      */
-    public function __construct(mixed $value);
+    public function __construct(
+        mixed $value
+    );
 }