+++ src/SingleParameterFactoryTrait.php
@@ -13,14 +13,16 @@ use Throwable;
 
 /**
  * @template TInput
+ * @phpstan-require-implements SingleParameterFactory<TInput>
  */
 trait SingleParameterFactoryTrait
 {
     /**
      * @param TInput|static $value
      */
-    public static function instance(mixed $value): static
-    {
+    public static function instance(
+        mixed $value
+    ): static {
         if ($value instanceof static) {
             return $value;
         }
@@ -31,8 +33,9 @@ trait SingleParameterFactoryTrait
     /**
      * @param TInput|static|null $value
      */
-    public static function orNull(mixed $value): ?static
-    {
+    public static function orNull(
+        mixed $value
+    ): ?static {
         if ($value === null) {
             return null;
         }