+++ src/Then.php
@@ -14,7 +14,9 @@ interface Then
     /**
      * @return $this
      */
-    public function then(callable $callback): Then;
+    public function then(
+        callable $callback
+    ): Then;
 
     /**
      * @param iterable<mixed, mixed> $values
@@ -31,7 +33,7 @@ interface Then
     public function thenIf(
         ?bool $truth,
         callable $yes,
-        callable $no = null
+        ?callable $no = null
     ): Then;
 
     /**
@@ -40,6 +42,6 @@ interface Then
     public function thenUnless(
         ?bool $truth,
         callable $no,
-        callable $yes = null
+        ?callable $yes = null
     ): Then;
 }