+++ src/ThenTrait.php
@@ -9,6 +9,9 @@ declare(strict_types=1);
 
 namespace DecodeLabs\Fluidity;
 
+/**
+ * @phpstan-require-implements Then
+ */
 trait ThenTrait
 {
     /**
@@ -16,8 +19,9 @@ trait ThenTrait
      *
      * @return $this
      */
-    public function then(callable $callback): Then
-    {
+    public function then(
+        callable $callback
+    ): Then {
         $callback($this);
 
         return $this;
@@ -49,7 +53,7 @@ trait ThenTrait
     public function thenIf(
         ?bool $truth,
         callable $yes,
-        callable $no = null
+        ?callable $no = null
     ): Then {
         if ($truth === true) {
             $yes($this, $truth);
@@ -69,7 +73,7 @@ trait ThenTrait
     public function thenUnless(
         ?bool $truth,
         callable $no,
-        callable $yes = null
+        ?callable $yes = null
     ): Then {
         if ($truth !== true) {
             $no($this, $truth);