+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## v0.3.5 (2025-02-07)
+* Fixed implicit nullable arguments
+* Added @phpstan-require-implements constraints
+* Made PHP8.1 minimum version
+
 ## v0.3.4 (2023-09-26)
 * Converted phpstan doc comments to generic
 * Migrated to use effigy in CI workflow